optimized the game by re-using blocks and enemies instead of deleting and creating them over and over again and fixed some warnings on the engine

+++ AstralEngine/src/AstralEngine/Core/Log.h
@@ -104,7 +104,7 @@ namespace AstralEngine
 			{
 				if (format[i] == '%')
 				{
-					float temp;
+					double temp;
 					bool correctFormat = true;
 					switch (format[i + 1])
 					{
